+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog] and this project adheres to
 - _All_ `Object::call` methods to be `unsafe` due to ability to break
   assumptions made by types such as `Array<A>`, where a type `B` can be inserted
   via `array.call_with("push", &[other_obj])`
+- `AnyObject::is_{nil|undefined|true|false}` to `const` functions
 
 ## [0.0.8] - 2019-05-23
 ### Added